拓片深度处理/requirements.txt
Pillow==10.0.0
numpy==1.26.0
scipy==1.11.3
opencv-python==4.8.0.76
